FLUSH_ICACHE defaults to __builtin___clear_cache if available

+++ arch/generic/machine.h
@@ -124,5 +124,5 @@ typedef void *Label;
 #define RELINFOBITS	8
 
 #ifdef HAVE___BUILTIN___CLEAR_CACHE
-#define FLUSH_ICACHE(addr,size) __builtin___clear_cache((void*)addr,(void*)addr+size)
+#define FLUSH_ICACHE(addr,size) __builtin___clear_cache((void*)addr,(void*)addr+(size_t)size)
 #endif
